# swagger/swagger.yml
openapi: 3.1.0
info:
  title: Java Spring Boot REST API
  version: 1.0.0
  description: Java Spring Boot REST API
servers:
  - url: http://localhost:8080/api/v1
paths:
  /product:
    post:
      summary: Create a product
      requestBody:
        required: true
        description: The product to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        200:
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products:
    get:
      summary: Get all products
      responses:
        200:
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /products/{id}:
    get:
      summary: Get a product by id
      parameters:
        - name: id
          in: path
          required: true
          description: The product id
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        404:  
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          description: The product id
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        description: The new product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        200:
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        404:  
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          description: The product id
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Product deleted successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Product deleted successfully"
components:
  schemas:
    ProductCreate:
      type: object
      properties:
        name:
          type: string
          example: "Sample Product"
        price:
          type: number
          format: double
          example: 9.99
        quantity:
          type: integer
          format: int32
          example: 10
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Sample Product"
        price:
          type: number
          format: double
          example: 9.99
        quantity:
          type: integer
          format: int32
          example: 10
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Product not found"